BambiCTF_2021/ringring/InvoiceApp/requirements.txt
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
configobj==5.0.6
docker==4.2.0
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Table==0.5.0
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-gnupg==0.4.6
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
systematic==4.8.7
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
gunicorn[eventlet]
psycopg2-binary==2.8.5
